backend/pyproject.toml
[project]
name = "assessmax-backend"
version = "0.1.0"
description = "Middle School Non-Academic Skills Measurement Engine - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "alembic>=1.13.0",
    "boto3>=1.34.0",
    "aiohttp>=3.9.0",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.8.0",
    "transformers>=4.38.0",
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "python-multipart>=0.0.9",
    "langdetect>=1.0.9",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
